Got persistence of AtomicRepository working.
It now saves itself as needed, particularly when the iCloud token is being saved, so that we can ensure there are no race conditions and data loss.

Sync actually does seem to work, but the app is very slow. We probably need some throttling of evens to make it better. I think we are merging into main on every keystroke.