Removed MergeableOrderedSet and added uniquing to MergeableArray.
Decided having an ordered set is confusing for most people. Better to work with standard arrays, and offer a simple way to remove duplicated elements (that are Identifiable). Calling this after a merge is a good idea for identifiable types. Perhaps will add this as option for merging in ForkedModel.