Fixed merging logic in CloudKitExchange.
Needed to add a new “restart” func.

We’ve moved to a single .cloudKit fork, together with an .uploadingToCloudKit branch that never merges but stages the upload while we wait for it to finish.

When the upload completes, we take the value from .uploadingToCloudKit, and we make that the new commonAncestor in .cloudKit, as we know that is the most recent version in the cloud. This is important, otherwise we can get changes being applied over and over causing them to appear in the multiple times.