Improved the data flow in Forkers
Removed onAppear.
Split the dual-purpose DetailView into Detail and Editor, removing a lot of ifs and complexity. Detail takes a Forker let and Editor takes a binding.
Consolidated all the states like editedForker, isEditing, hasAppeared, showingAddForker into a single EditorConfig source of truth for adding and editing.
Different toolbars for the add new sheet and Detail editing.
Removed Group which can be error prone with modifiers and used an overlay for the ContentUnavailableView.