Fixed recursive merge of ArrayOfIdentifiable and Dictionary.
When the values are Mergeable, they should be recursively merged, but they were not being due to issues calling the wrong merge funcs related to type erasure with the `Merger` protocol.

This has been cleaned up, and quite a detailed test added that includes recursive merging of both dictionaries and arrays of identifiables.