Better handling of CloudKitExchange assets during upload.
Previously, we were creating a temporary file, and cleaning it up after a time period. This is hacky, and often CKSyncEngine would not find the file anymore and have to schedule a new upload.

Now we have a dedicated temp dir, and clean up  in init and when the upload succeeds.