Removed the dependency of ForkedModelMacros on other modules (eg ForkedMerge).
This caused SPM to fail in Xcode release builds with multiple archs (eg macOS). It was working with xcodebuild, so seems it is probably a bug in Xcode itself, but this change simplifies the dependencies.

The error itself was that there were two rules to create the library Forked.o and ForkedMerge.o. This was because they were both used in the macros.

To fix this, we no longer have a dependence on them in the macros. To achieve this, we have duplicated the enums. Apparently this is acceptable, because the macros are used in a different context, and just produce code.